Upgrade to ruby 2.1.8

+++ Rakefile
@@ -4,10 +4,13 @@
 require File.expand_path('../config/application', __FILE__)
 
 require 'fileutils'
-Rake::Task["assets:clean"].enhance do
-  puts 'my assets:precompile hook has started'
-  FileUtils.remove_dir "#{Rails.root}/public/assets", true
-  puts 'my assets:precompile hook is finished'
+
+if Rails.env.production?
+  Rake::Task["assets:clean"].enhance do
+    puts 'my assets:precompile hook has started'
+    FileUtils.remove_dir "#{Rails.root}/public/assets", true
+    puts 'my assets:precompile hook is finished'
+  end
 end
 
 Americo::Application.load_tasks
